src/pyproject.toml
[tool.poetry]
name = "StudyCobra"
version = "0.1.0"
description = ""
authors = ["ivan@simantiev.com"]

[tool.poetry.dev-dependencies]
mypy = "*"
black = "==20.8b1"
isort = "==5.7.0"
pytest = "^6.2.1"
pytest-env = "^0.6.2"
ipython = "==7.19.0"
flake8 = "==3.8.4"
coverage = "==5.3.1"
flake8-logging-format = "*"
flake8-alfred = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-coding = "*"
flake8-commas = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-eradicate = "*"
flake8-multiline-containers = "*"
flake8-pep3101 = "*"
flake8-print = "*"
flake8-pytest-style = "*"
flake8-return = "*"
flake8-quotes = "*"
flake8-simplify = "*"
flake8-sql = "*"
pep8-naming = "*"
flake8-cognitive-complexity = "*"
flake8-isort = "*"
sqlalchemy-stubs = "^0.4"
flake8-annotations = "^2.6.2"
types-requests = "^2.25.6"
types-aiofiles = "^0.8.4"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.0"
pydantic = "==1.7.3"
uvicorn = "^0.13.4"
requests = "==2.25.1"
sentry-sdk = "==1.3.1"
psutil = "==5.8.0"
httpx = "^0.17.0"
SQLAlchemy = "^1.3.23"
asyncpg = "^0.24.0"
databases = "^0.4.2"
psycopg2-binary = "^2.8.6"
aio-pika = "^6.8.0"
structlog = "^21.1.0"
alembic = "^1.5.7"
aiofiles = "^0.6.0"
PyJWT = "^2.0.1"
gunicorn = "^20.1.0"
structlog-sentry = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
env = "TESTING=True"


[tool.black]
line-length = 100
py38 = true
include = '\.py$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | env
  | test-reports
  | ptvs_virtualenv_proxy.py
)/
'''

[tool.isort]
line_length = 100
default_section = "FIRSTPARTY"
force_single_line = "True"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_local_folder = "api,core,crud,db,enums,models,schemas,sdk,tests,validators"
